Move get_iap_garage funciton to common utils
