+++ iap/utils.py
@@ -1,17 +1,10 @@
 import datetime
-import os
-from typing import List, Optional, Dict
 
-from gql.dsl import dsl_gql, DSLQuery
-from sqlalchemy import func, distinct, select, Date, cast
+from sqlalchemy import func, Date, cast
 
 from common import logger
-from common._crypto import Account
-from common._graphql import GQL
 from common.enums import ReceiptStatus
-from common.models.product import FungibleItemProduct
 from common.models.receipt import Receipt
-from common.utils import fetch_kms_key_id
 
 
 def get_purchase_count(sess, agent_addr: str, product_id: int, hour_limit: int) -> int:
@@ -26,7 +19,7 @@ def get_purchase_count(sess, agent_addr: str, product_id: int, hour_limit: int) 
     """
     # NOTE: Subtract 24 hours from incoming hour_limit.
     #  Because last 24 hours means today. Using `datetime.date()` function, timedelta -24 hours makes yesterday.
-    start = (datetime.datetime.utcnow() - datetime.timedelta(hours=hour_limit-24)).date()
+    start = (datetime.datetime.utcnow() - datetime.timedelta(hours=hour_limit - 24)).date()
     purchase_count = (
         sess.query(func.count(Receipt.id)).filter_by(product_id=product_id, agent_addr=agent_addr)
         .filter(Receipt.status.in_(
@@ -38,40 +31,3 @@ def get_purchase_count(sess, agent_addr: str, product_id: int, hour_limit: int) 
         f"Agent {agent_addr} purchased product {product_id} {purchase_count} times in {hour_limit} hours from {start}"
     )
     return purchase_count
-
-
-def get_iap_garage(sess) -> List[Optional[Dict]]:
-    """
-    Get NCG balance and fungible item count of IAP address.
-    :return:
-    """
-    stage = os.environ.get("STAGE", "development")
-    region_name = os.environ.get("REGION_NAME", "us-east-2")
-    client = GQL()
-    account = Account(fetch_kms_key_id(stage, region_name))
-
-    fungible_id_list = sess.scalars(select(distinct(FungibleItemProduct.fungible_item_id))).fetchall()
-
-    query = dsl_gql(
-        DSLQuery(
-            client.ds.StandaloneQuery.stateQuery.select(
-                client.ds.StateQuery.garages.args(
-                    agentAddr=account.address,
-                    fungibleItemIds=fungible_id_list,
-                ).select(
-                    client.ds.GaragesType.agentAddr,
-                    client.ds.GaragesType.fungibleItemGarages.select(
-                        client.ds.FungibleItemGarageWithAddressType.fungibleItemId,
-                        client.ds.FungibleItemGarageWithAddressType.count,
-                    )
-                )
-            )
-        )
-    )
-    resp = client.execute(query)
-    if "errors" in resp:
-        msg = f"GQL failed to get IAP garage: {resp['errors']}"
-        logger.error(msg)
-        raise Exception(msg)
-
-    return resp["stateQuery"]["garages"]["fungibleItemGarages"]
